// lib/types.ts
export const transmissionTypes = ["youtube", "tv", "pauta", "viagem"] as const;
export type TransmissionType = (typeof transmissionTypes)[number];

export type EventStatus = "Agendado" | "Concluído" | "Cancelado";
export type EventTurn = "Manhã" | "Tarde" | "Noite";

export interface Event {
  id: string;
  name: string;
  date: Date;
  location: string;
  transmission: TransmissionType[];
  pauta?: string;
  color: string;
  transmissionOperator?: string;
  cinematographicReporter?: string;
  reporter?: string;
  producer?: string;
  status: EventStatus;
  turn: EventTurn;
  departure?: Date | null;
  arrival?: Date | null;
}

export type EventFormData = Omit<Event, "id" | "color" | "status" | "turn">;
// ...
export interface RepeatSettings {
    frequency: 'daily' | 'weekly' | 'monthly';
    count: number;
}